# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Postmark Service API",
    "description": "Email sending and tracking service built on Postmark. Handles email delivery via the broadcast message stream, webhook processing for delivery events, and integrates with a runs-service for cost tracking.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://postmark.mcpfactory.org",
      "description": "Production"
    },
    {
      "url": "http://localhost:3010",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Service-to-service API key"
      }
    },
    "schemas": {
...
      "SendEmailResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "messageId": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "sendingId": {
            "type": "string"
          },
          "errorCode": {
            "type": "number"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "required": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "InsufficientCreditsResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Insufficient credits"
            ]
          },
          "balance_cents": {
            "type": "number",
            "nullable": true
          },
          "required_cents": {
            "type": "number"
          }
        },
        "required": [
          "error",
          "balance_cents",
          "required_cents"
        ]
      },
      "SendEmailRequest": {
        "type": "object",
        "properties": {
          "brandId": {
            "type": "string",
            "description": "Brand ID"
          },
          "campaignId": {
            "type": "string",
            "description": "Campaign ID"
          },
          "featureSlug": {
            "type": "string",
            "description": "Feature slug for tracking"
          },
          "workflowSlug": {
            "type": "string",
            "description": "Workflow slug for tracking/grouping"
          },
          "leadId": {
            "type": "string",
            "description": "Lead ID for tracking and dedup"
          },
          "from": {
            "type": "string",
            "description": "Sender email address. If omitted, resolved from key-service (provider: postmark-from-address)."
          },
          "to": {
            "type": "string",
            "description": "Recipient email address"
          },
          "subject": {
            "type": "string",
            "description": "Email subject line"
          },
          "htmlBody": {
            "type": "string",
            "description": "HTML email body"
          },
          "textBody": {
            "type": "string",
            "description": "Plain text email body"
          },
          "cc": {
            "type": "string"
          },
          "bcc": {
            "type": "string"
          },
          "replyTo": {
            "type": "string"
          },
          "tag": {
            "type": "string"
          },
          "headers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "value": {
                  "type": "string"
                }
              },
              "required": [
                "name",
                "value"
              ]
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "trackOpens": {
            "type": "boolean",
            "default": true
          },
          "trackLinks": {
            "type": "string",
            "enum": [
              "None",
              "HtmlAndText",
              "HtmlOnly",
              "TextOnly"
            ]
          }
        },
        "required": [
          "to",
          "subject"
        ]
      },
      "BatchSendResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "successCount": {
            "type": "number"
          },
          "failCount": {
            "type": "number"
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "to": {
                  "type": "string"
                },
                "success": {
                  "type": "boolean"
                },
                "messageId": {
                  "type": "string"
                },
                "sendingId": {
                  "type": "string"
                },
                "errorCode": {
                  "type": "number"
                },
                "message": {
                  "type": "string"
                },
                "error": {
                  "type": "string"
                }
              },
              "required": [
                "to",
                "success"
              ]
            }
          }
        },
        "required": [
          "total",
          "successCount",
          "failCount",
          "results"
        ]
      },
      "BatchSendRequest": {
        "type": "object",
        "properties": {
          "emails": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "brandId": {
                  "type": "string"
                },
                "campaignId": {
                  "type": "string"
                },
                "featureSlug": {
                  "type": "string"
                },
                "workflowSlug": {
                  "type": "string"
                },
                "leadId": {
                  "type": "string"
                },
                "from": {
                  "type": "string"
                },
                "to": {
                  "type": "string"
                },
                "subject": {
                  "type": "string"
                },
                "htmlBody": {
                  "type": "string"
                },
                "textBody": {
                  "type": "string"
                },
                "cc": {
                  "type": "string"
                },
                "bcc": {
                  "type": "string"
                },
                "replyTo": {
                  "type": "string"
                },
                "tag": {
                  "type": "string"
                },
                "headers": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "value": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "name",
                      "value"
                    ]
                  }
                },
                "metadata": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                },
                "trackOpens": {
                  "type": "boolean",
                  "default": true
                },
                "trackLinks": {
                  "type": "string",
                  "enum": [
                    "None",
                    "HtmlAndText",
                    "HtmlOnly",
                    "TextOnly"
                  ]
                }
              },
              "required": [
                "to",
                "subject"
              ]
            },
            "minItems": 1,
            "maxItems": 500
          }
        },
        "required": [
          "emails"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/send": {
      "post": {
        "summary": "Send a single email",
        "description": "Send an email via Postmark and record it in the database. Runs-service integration is BLOCKING.",
        "tags": [
          "Email Sending"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s), comma-separated for multi-brand campaigns (injected by workflow-service). Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s), comma-separated for multi-brand campaigns (injected by workflow-service). Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug (injected by workflow-service)"
            },
            "required": false,
            "description": "Feature slug (injected by workflow-service)",
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow slug (injected by workflow-service)",
            "name": "x-workflow-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendEmailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Insufficient credits (platform key only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientCreditsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/send/batch": {
      "post": {
        "summary": "Send batch emails",
        "description": "Send up to 500 emails in one request. Runs-service integration is BLOCKING for each email.",
        "tags": [
          "Email Sending"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Campaign ID (injected by workflow-service)"
            },
            "required": false,
            "description": "Campaign ID (injected by workflow-service)",
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Brand ID(s), comma-separated for multi-brand campaigns (injected by workflow-service). Example: uuid1,uuid2,uuid3"
            },
            "required": false,
            "description": "Brand ID(s), comma-separated for multi-brand campaigns (injected by workflow-service). Example: uuid1,uuid2,uuid3",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Feature slug (injected by workflow-service)"
            },
            "required": false,
            "description": "Feature slug (injected by workflow-service)",
            "name": "x-feature-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Workflow slug (injected by workflow-service)"
            },
            "required": false,
            "description": "Workflow slug (injected by workflow-service)",
            "name": "x-workflow-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchSendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSendResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Insufficient credits (platform key only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientCreditsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}